Clean up styling and add error message for server failure
Some `<div>`s were replaced and removed to clean up the styling of the
main page.

Added a `setError()` message for when the server fails to respond or
fails to return an answer value back to the user. This could still be
made more robust and give more helpful error messages such as when the
server OOMs or other pytorch events fail in AllenNLP or the server is
not responding.